+++ ATQ1MR_HFT_2021221.Logic/Services/ProcessorLogic.cs
@@ -72,7 +72,15 @@ namespace ATQ1MR_HFT_2021221.Logic.Services
         }
         public void Delete(int id)
         {
-            _processorRepository.Delete(id);
+            var v = _processorRepository.Read(id);
+            if (v != null)
+            {
+                _processorRepository.Delete(id);
+            }
+            else
+            {
+                throw new Exception("No entity found!");
+            }
         }
     }
 }
